requirements.txt
Flask==3.0.2
pymongo==4.6.3
requests==2.31.0
python-dotenv==1.0.1
sqlalchemy==2.0.29